pyproject.toml
[project]
name = "lenny-clone"
version = "0.1.0"
description = "RAG-powered Lenny Rachitsky clone"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.2",
    "openai>=1.12.0",
    "youtube-transcript-api>=0.6.1",
    "apify-client>=1.6.0",
    "langchain>=0.1.0",
    "langchain-text-splitters>=0.0.1",
    "streamlit>=1.31.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "feedparser>=6.0.10",
    "rich>=13.7.0",
    "chromadb>=1.3.5",
    "yt-dlp>=2025.11.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ingestion", "agent", "app", "evaluation"]
